Rename "yourproject" to "webgame"
