+++ webgame-ml/README.md
@@ -6,7 +6,7 @@ A standard template I use for my RL projects.
 
 First, edit `pyproject.toml` and the folder names to your project. There are
 also some import statements you'll have to change, but the easiest way to do
-that is a global replacement of `yourproject-ml` with your new project name.
+that is a global replacement of `webgame-ml` with your new project name.
 
 Next, install the project. You'll need Poetry and >=Python3.9.
