+++ webgame-ml/webgame/experiments/test_rust.py
@@ -14,11 +14,11 @@ from gymnasium.vector import SyncVectorEnv
 from torch.distributions import Categorical
 from tqdm import tqdm
 
-from yourproject.algorithms.ppo import train_ppo
-from yourproject.algorithms.rollout_buffer import RolloutBuffer
-from yourproject.conf import entity
-from yourproject.utils import init_orthogonal
-from yourproject_rust import CartpoleEnv  # type: ignore
+from webgame.algorithms.ppo import train_ppo
+from webgame.algorithms.rollout_buffer import RolloutBuffer
+from webgame.conf import entity
+from webgame.utils import init_orthogonal
+from webgame_rust import CartpoleEnv  # type: ignore
 
 _: Any
