+++ webgame-ml/webgame_rust/Cargo.toml
@@ -1,15 +1,15 @@
 [package]
-name = "yourproject_rust"
+name = "webgame_rust"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
-name = "yourproject_rust"
+name = "webgame_rust"
 crate-type = ["cdylib"]
 
 [dependencies]
 pyo3 = "0.18.3"
 rand = "0.8.5"
 bevy = "0.12.0"
-yourproject-game = { path = "../../yourproject-game" }
+webgame-game = { path = "../../webgame-game" }
