+++ webgame-ml/webgame_rust/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["maturin>=0.15,<0.16"]
 build-backend = "maturin"
 
 [project]
-name = "yourproject_rust"
+name = "webgame_rust"
 requires-python = ">=3.7"
 classifiers = [
     "Programming Language :: Rust",
